Cleaned up the source a bit. In the process I changed the way user interface structure is passed to the Application.
Basically it works so now that you need to put all of your user interface structures below a class (I have used class named UIStructure but the class name does not make any big difference). It is important that the root structure of the user interface is put in a string named "root_structure". Substructures can be named as you like as long as the name matches to a structure found in the class.

Note that I updated the examples to reflect the current way of passing structure to application.