Implemented part of the state event backend. I also refactored BaseEventManager and separated ElementEventContainer and EventContainer out of it. It may be that those two classes will be merged at some later time.
State event architecture is still missing the part that actually checks if coordinate is inside given element. For this I need to implement some sort of system that initializes x and y coordinates of elements explicitly. Currently this is embedded in the container rendering functions. Probably a good workaround would be to make render functions of all objects to store x, y when they are rendered and then execute rendering before testing.