Made state events work in Blender. Note that they have to be explicitly enabled by using USE_STATE_EVENTS flag like: import bui.blender.application bui.blender.application.USE_STATE_EVENTS = True
The reason for this is that current implementation is somewhat slow. This is due to the fact that it forces redraw each time mouse location changes. There might be some nicer way to implement this so that redraw is called only when necessary.

image_state_event.py shows a small example of how to use this new feature. If you move mouse on top of the image of the mandril, it will generate new images on the container below.